requirements.txt
alabaster==1.0.0
astroid==3.3.11
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dill==0.4.0
docutils==0.21.2
ghp-import==2.1.0
gitdb==4.0.12
gitpython==3.1.45
html5lib==1.1
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
isort==6.0.1
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
logging-tree==1.10
lxml==6.0.0
mako==1.3.10
markdown==3.8.2
markupsafe==3.0.2
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-rss-plugin==1.17.1
mkdocs-get-deps==0.2.0
mypy==1.17.1
mypy-extensions==1.1.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pyclassifiers==0.0.20
pycmdtools==0.0.85
pydmt==0.3.32
pyfakeuse==0.0.9
pygments==2.19.2
pylint==3.3.8
pylogconf==0.0.38
pymakehelper==0.0.37
pytconf==0.1.19
pytest==8.4.1
python-dateutil==2.9.0.post0
pytidylib==0.3.2
pytz==2025.2
pyyaml==6.0.2
pyyaml-env-tag==1.1
referencing==0.36.2
requests==2.32.5
roman-numerals-py==3.1.0
rpds-py==0.27.1
ruff==0.12.8
six==1.17.0
smmap==5.0.2
snowballstemmer==3.0.1
soupsieve==2.7
sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
termcolor==3.1.0
tomlkit==0.13.3
tqdm==4.67.1
types-pyyaml==6.0.12.20250809
types-requests==2.32.4.20250809
types-termcolor==1.1.6.2
typing-extensions==4.15.0
tzdata==2025.2
unidecode==1.4.0
urllib3==2.5.0
venv-run==0.2.0
watchdog==6.0.0
webencodings==0.5.1
yattag==1.16.1
